+++ src/gax.ts
@@ -18,6 +18,8 @@
  * Google API Extensions
  */
 
+import type {Message} from 'protobufjs';
+import {warn} from './warnings';
 import {BundleOptions} from './bundlingCalls/bundleExecutor';
 import {toLowerCamelCase} from './util';
 
@@ -701,4 +703,21 @@ export function constructSettings(
   }
 
   return defaults;
+}
+
+export function createByteLengthFunction(message: typeof Message) {
+  return function getByteLength(obj: {}) {
+    try {
+      return message.encode(obj).finish().length;
+    } catch (err) {
+      const stringified = JSON.stringify(obj);
+      warn(
+        'error_encoding_protobufjs_object',
+        `Cannot encode protobuf.js object: ${stringified}: ${err}`
+      );
+      // We failed to encode the object properly, let's just return an upper boundary of its length.
+      // It's only needed for calculating the size of the batch, so it's safe if it's bigger than needed.
+      return stringified.length;
+    }
+  };
 }