+++ test/unit/compileProtos.ts
@@ -70,7 +70,11 @@ describe('compileProtos tool', () => {
     assert(
       js.toString().includes('http://www.apache.org/licenses/LICENSE-2.0')
     );
-    assert(js.toString().includes('require("google-gax").protobufMinimal'));
+    assert(
+      js
+        .toString()
+        .includes('require("google-gax/build/src/protobuf").protobufMinimal')
+    );
     assert(!js.toString().includes('require("protobufjs/minimal")'));
 
     // check that it uses proper root object; it's taken from fixtures/package.json
@@ -79,12 +83,15 @@ describe('compileProtos tool', () => {
     const ts = await readFile(expectedTSResultFile);
     assert(ts.toString().includes('TestMessage'));
     assert(ts.toString().includes('LibraryService'));
-    assert(ts.toString().includes('import * as Long'));
+    assert(ts.toString().includes('import Long = require'));
+    assert(!ts.toString().includes('import * as Long'));
     assert(
       ts.toString().includes('http://www.apache.org/licenses/LICENSE-2.0')
     );
     assert(
-      ts.toString().includes('import {protobuf as $protobuf} from "google-gax"')
+      ts
+        .toString()
+        .includes('import type {protobuf as $protobuf} from "google-gax"')
     );
     assert(!ts.toString().includes('import * as $protobuf from "protobufjs"'));
   });
@@ -105,7 +112,11 @@ describe('compileProtos tool', () => {
     assert(
       js.toString().includes('http://www.apache.org/licenses/LICENSE-2.0')
     );
-    assert(js.toString().includes('require("google-gax").protobufMinimal'));
+    assert(
+      js
+        .toString()
+        .includes('require("google-gax/build/src/protobuf").protobufMinimal')
+    );
     assert(!js.toString().includes('require("protobufjs/minimal")'));
 
     // check that it uses proper root object; it's taken from fixtures/package.json
@@ -114,12 +125,15 @@ describe('compileProtos tool', () => {
     const ts = await readFile(expectedTSResultFile);
     assert(ts.toString().includes('TestMessage'));
     assert(ts.toString().includes('LibraryService'));
-    assert(ts.toString().includes('import * as Long'));
+    assert(ts.toString().includes('import Long = require'));
+    assert(!ts.toString().includes('import * as Long'));
     assert(
       ts.toString().includes('http://www.apache.org/licenses/LICENSE-2.0')
     );
     assert(
-      ts.toString().includes('import {protobuf as $protobuf} from "google-gax"')
+      ts
+        .toString()
+        .includes('import type {protobuf as $protobuf} from "google-gax"')
     );
     assert(!ts.toString().includes('import * as $protobuf from "protobufjs"'));
   });
@@ -146,7 +160,8 @@ describe('compileProtos tool', () => {
     assert(fs.existsSync(expectedTSResultFile));
     const ts = await readFile(expectedTSResultFile);
 
-    assert(ts.toString().includes('import * as Long'));
+    assert(ts.toString().includes('import Long = require'));
+    assert(!ts.toString().includes('import * as Long'));
     assert(
       ts.toString().includes('http://www.apache.org/licenses/LICENSE-2.0')
     );