+++ test/unit/transcoding.ts
@@ -32,6 +32,7 @@ import {
   buildQueryStringComponents,
   requestChangeCaseAndCleanup,
   overrideHttpRules,
+  TranscodedRequest,
 } from '../../src/transcoding';
 import * as assert from 'assert';
 import {
@@ -40,7 +41,7 @@ import {
 } from '../../src/util';
 import * as protobuf from 'protobufjs';
 import {testMessageJson} from '../fixtures/fallbackOptional';
-import {echoProtoJson} from '../fixtures/echoProtoJson';
+import echoProtoJson = require('../fixtures/echo.json');
 import {google} from '../../protos/http';
 
 describe('gRPC to HTTP transcoding', () => {
@@ -192,6 +193,30 @@ describe('gRPC to HTTP transcoding', () => {
       transcode({unknownField: 'project'}, parsedOptions),
       undefined
     );
+  });
+
+  it('should not change user inputted fields to camel case', () => {
+    const root = protobuf.Root.fromJSON(testMessageJson);
+    const testMessageFields = root.lookupType('TestMessage').fields;
+    const request: RequestType = {
+      projectId: 'test-project',
+      content: 'test-content',
+      labels: {'i-am-vm': 'true'},
+    };
+    const parsedOptions: ParsedOptionsType = [
+      {
+        '(google.api.http)': {
+          post: 'projects/{project_id}',
+          body: '*',
+        },
+      },
+    ];
+    const transcoded = transcode(request, parsedOptions, testMessageFields);
+    assert.deepStrictEqual(transcoded?.url, 'projects/test-project');
+    assert.deepStrictEqual(transcoded?.data, {
+      content: 'test-content',
+      labels: {'i-am-vm': 'true'},
+    });
   });
 
   it('transcode should not decapitalize the first capital letter', () => {
@@ -588,26 +613,17 @@ describe('validate proto3 field with default value', () => {
           body: '*',
         },
       },
-      {
-        '(google.api.method_signature)': 'project_id, content',
-      },
     ];
     const transcoded = transcode(request, parsedOptions, badTestMessageFields);
     assert.deepStrictEqual(
-      transcoded?.url,
+      (transcoded as TranscodedRequest)?.url,
       'projects/test-project/contents/test-content'
     );
   });
-  it('should throw error if required field has not been setted', () => {
-    const requests: RequestType[] = [
-      {
-        projectId: 'test-project',
-        content: 'undefined',
-      },
-      {
-        projectId: 'test-project',
-      },
-    ];
+  it('should throw error if required field has not been set', () => {
+    const request: RequestType = {
+      projectId: 'test-project',
+    };
     const parsedOptions: ParsedOptionsType = [
       {
         '(google.api.http)': {
@@ -615,16 +631,11 @@ describe('validate proto3 field with default value', () => {
           body: '*',
         },
       },
-      {
-        '(google.api.method_signature)': 'project_id, content',
-      },
     ];
-    for (const request of requests) {
-      assert.throws(
-        () => transcode(request, parsedOptions, testMessageFields),
-        Error
-      );
-    }
+    assert.throws(
+      () => transcode(request, parsedOptions, testMessageFields),
+      /Error: Required field content is not present in the request/
+    );
   });
   it('when body="*", all required field should emitted in body', () => {
     const request: RequestType = {
@@ -640,21 +651,19 @@ describe('validate proto3 field with default value', () => {
       },
     ];
     const transcoded = transcode(request, parsedOptions, testMessageFields);
-    assert.deepStrictEqual(transcoded?.url, 'projects/test-project');
-    assert.deepStrictEqual(transcoded?.data, {content: 'test-content'});
+    assert.deepStrictEqual(
+      (transcoded as TranscodedRequest)?.url,
+      'projects/test-project'
+    );
+    assert.deepStrictEqual((transcoded as TranscodedRequest)?.data, {
+      content: 'test-content',
+    });
   });
   it('when body="*", unset optional field should remove from body', () => {
-    const requests: RequestType[] = [
-      {
-        projectId: 'test-project',
-        content: 'test-content',
-        optionalValue: 'undefined',
-      },
-      {
-        projectId: 'test-project',
-        content: 'test-content',
-      },
-    ];
+    const request: RequestType = {
+      projectId: 'test-project',
+      content: 'test-content',
+    };
     const parsedOptions: ParsedOptionsType = [
       {
         '(google.api.http)': {
@@ -663,27 +672,18 @@ describe('validate proto3 field with default value', () => {
         },
       },
     ];
-    for (const request of requests) {
-      const transcoded = transcode(request, parsedOptions, testMessageFields);
-      assert.deepStrictEqual(
-        transcoded?.url,
-        'projects/test-project/contents/test-content'
-      );
-      assert.deepStrictEqual(transcoded?.data, {});
-    }
+    const transcoded = transcode(request, parsedOptions, testMessageFields);
+    assert.deepStrictEqual(
+      (transcoded as TranscodedRequest)?.url,
+      'projects/test-project/contents/test-content'
+    );
+    assert.deepStrictEqual((transcoded as TranscodedRequest)?.data, {});
   });
   it('unset optional fields should not appear in query params', () => {
-    const requests: RequestType[] = [
-      {
-        projectId: 'test-project',
-        content: 'test-content',
-        optionalValue: 'undefined',
-      },
-      {
-        projectId: 'test-project',
-        content: 'test-content',
-      },
-    ];
+    const request: RequestType = {
+      projectId: 'test-project',
+      content: 'test-content',
+    };
     const parsedOptions: ParsedOptionsType = [
       {
         '(google.api.http)': {
@@ -695,12 +695,16 @@ describe('validate proto3 field with default value', () => {
         '(google.api.method_signature)': 'project_id, content',
       },
     ];
-    for (const request of requests) {
-      const transcoded = transcode(request, parsedOptions, testMessageFields);
-      assert.deepStrictEqual(transcoded?.url, 'projects/test-project');
-      assert.deepStrictEqual(transcoded?.data, 'test-content');
-      assert.deepStrictEqual(transcoded.queryString, '');
-    }
+    const transcoded = transcode(request, parsedOptions, testMessageFields);
+    assert.deepStrictEqual(
+      (transcoded as TranscodedRequest)?.url,
+      'projects/test-project'
+    );
+    assert.deepStrictEqual(
+      (transcoded as TranscodedRequest)?.data,
+      'test-content'
+    );
+    assert.deepStrictEqual((transcoded as TranscodedRequest).queryString, '');
   });
 });